+++ R/test.R
@@ -207,7 +207,8 @@ test.emmGrid = function(object, null = 0,
 #' whereas type III tests are tests of model coefficients -- which may or may
 #' not have anything to do with EMMs or contrasts.
 #' 
-#' @param object a fitted model or an \code{emmGrid} 
+#' @param object a fitted model, \code{emmGrid}, or \code{emm_list}. If the
+#'   latter, its first element is used.
 #' @param cov.reduce a function.
 #'    If \code{object} is a fitted model, it is
 #'    replaced by \code{ref_grid(object, cov.reduce = cov.reduce, ...)}.